pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "graff-experiments"
version = "0.1.0"
description = "Research code for GRAFF fine-tuning experiments"
requires-python = ">=3.9"

# Runtime dependencies
dependencies = [
  "torch>=2.0.0",
  "transformers>=4.39.0",
  "datasets>=2.18.0",
  "peft>=0.7.0",
  "hydra-core>=1.3.2",
  "wandb>=0.16.0",
  "optuna>=3.5.0",
  "networkx>=3.2.0",
  "scikit-learn>=1.3.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "pandas>=2.1.0",
  "numpy>=1.24.0",
  "scipy>=1.11.0"
]
